backend/requirements.txt
# Django Core
Django==5.2.8
djangorestframework==3.15.2

# Database
mysqlclient==2.2.7

# CORS
django-cors-headers==4.6.0

# Image handling
Pillow==11.0.0

# Utilities
python-dateutil==2.9.0
